[UI] Phase 4: Consolidate workspace UI under ui/components/workspaces
Closes #18746

Moves the two workspace homes into a single canonical location:

- ui/components/SpacesSwitcher/* -> ui/components/workspaces/SpacesSwitcher/*
- ui/components/Lifecycle/Workspaces/* -> ui/components/workspaces/*

Imports updated in Header.tsx, ViewInfoModal.tsx, and Lifecycle/index.tsx.
Internal relative imports inside moved files rewritten to use the @/ alias
so they no longer depend on previous depth.

eslint.config.js path allow-lists updated to reflect new locations.

Phase 4 parent: #18659. Pairs with #18750 (environments) in Lifecycle/;
different children so safe to land in parallel.
