+++ ui/components/Lifecycle/index.tsx
@@ -1,4 +1,4 @@
-import WorkspacesComponent from './Workspaces';
+import WorkspacesComponent from '../workspaces';
 import EnvironmentComponent from './Environments';
 
 export { WorkspacesComponent, EnvironmentComponent };
