+++ ui/components/workspaces/SpacesSwitcher/SharedContent.tsx
