+++ ui/components/workspaces/SpacesSwitcher/UserAvatarComponent.tsx
