+++ ui/components/workspaces/SpacesSwitcher/WorkspaceContent.tsx
@@ -31,7 +31,7 @@ import {
   useAssignViewToWorkspaceMutation,
 } from '@/rtk-query/workspace';
 import { getDefaultFilterType, useContentDelete, useContentDownload } from './hooks';
-import ExportModal from '../ExportModal';
+import ExportModal from '@/components/ExportModal';
 import { WorkspaceModalContext } from '@/utils/context/WorkspaceModalContextProvider';
 import { useRouter } from 'next/router';
 import { useSelector } from 'react-redux';
