+++ ui/components/workspaces/SpacesSwitcher/WorkspaceModal.tsx
@@ -21,12 +21,12 @@ import {
   useMediaQuery,
   useTheme,
 } from '@sistent/sistent';
-import { WorkspacesComponent } from '../Lifecycle';
+import { WorkspacesComponent } from '@/components/Lifecycle';
 import { iconMedium, iconSmall } from 'css/icons.styles';
 import MyViewsContent from './MyViewsContent';
 import MyDesignsContent from './MyDesignsContent';
 import RecentContent from './RecentContent';
-import { useGetWorkspacesQuery } from '../../rtk-query/workspace';
+import { useGetWorkspacesQuery } from '@/rtk-query/workspace';
 import { DrawerHeader, StyledDrawer, StyledMainContent, StyledModal } from './styles';
 import WorkspaceContent from './WorkspaceContent';
 import { useGetProviderCapabilitiesQuery, useGetSelectedOrganization } from '@/rtk-query/user';
