+++ ui/components/workspaces/styles.tsx
