+++ ui/eslint.config.js
@@ -42,7 +42,7 @@ const legacyRestrictedImportOffenders = [
   'components/Header.tsx',
   'components/Lifecycle/Environments/environment-card.tsx',
   'components/Lifecycle/Environments/index.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceGridView.tsx',
+  'components/workspaces/WorkspaceGridView.tsx',
   'components/MesheryAdapterPlayComponent.tsx',
   'components/MesheryChart.tsx',
   'components/MesheryMeshInterface/PatternService/RJSF.tsx',
@@ -58,9 +58,9 @@ const legacyRestrictedImportOffenders = [
   'components/ReactSelectWrapper.tsx',
   'components/RelationshipBuilder/RelationshipFormStepper.tsx',
   'components/Settings/Registry/Stepper/CSVStepper.tsx',
-  'components/SpacesSwitcher/MainDesignsContent.tsx',
-  'components/SpacesSwitcher/WorkspaceModal.tsx',
-  'components/SpacesSwitcher/components.tsx',
+  'components/workspaces/SpacesSwitcher/MainDesignsContent.tsx',
+  'components/workspaces/SpacesSwitcher/WorkspaceModal.tsx',
+  'components/workspaces/SpacesSwitcher/components.tsx',
   'components/UserPreferences/index.tsx',
   'components/connections/ConnectionChip.tsx',
   'components/connections/ConnectionTable.tsx',
@@ -88,7 +88,7 @@ const legacyLiteralColorOffenders = [
   'components/Lifecycle/General/empty-state/curvedArrowIcon.tsx',
   'components/Lifecycle/General/empty-state/index.tsx',
   'components/Lifecycle/General/flip-card/index.tsx',
-  'components/Lifecycle/Workspaces/index.tsx',
+  'components/workspaces/index.tsx',
   'components/LoadingComponents/Animations/AnimatedFilter.tsx',
   'components/LoadingComponents/Animations/AnimatedLightMeshery.tsx',
   'components/LoadingComponents/Animations/AnimatedMeshPattern.tsx',
@@ -123,8 +123,8 @@ const legacyLiteralColorOffenders = [
   'components/Settings/MesherySettings.tsx',
   'components/Settings/Registry/MeshModelDetails.tsx',
   'components/Settings/Registry/helper.ts',
-  'components/SpacesSwitcher/WorkspaceSwitcher.tsx',
-  'components/SpacesSwitcher/styles.tsx',
+  'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
+  'components/workspaces/SpacesSwitcher/styles.tsx',
   'components/TypingFilter/style.tsx',
   'components/UserPreferences/index.tsx',
   'components/UserPreferences/style.tsx',
@@ -228,10 +228,10 @@ const legacyInlineStyleOffenders = [
   'components/Header.tsx',
   'components/HeaderMenu.tsx',
   'components/Lifecycle/General/empty-state/index.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceActionList.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceDataTable.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceGridView.tsx',
-  'components/Lifecycle/Workspaces/index.tsx',
+  'components/workspaces/WorkspaceActionList.tsx',
+  'components/workspaces/WorkspaceDataTable.tsx',
+  'components/workspaces/WorkspaceGridView.tsx',
+  'components/workspaces/index.tsx',
   'components/LoadingComponents/Animations/AnimatedMeshSync.tsx',
   'components/LoadingComponents/LoadingComponent.tsx',
   'components/LoadingComponents/LoadingComponentServer.tsx',
@@ -302,20 +302,20 @@ const legacyInlineStyleOffenders = [
   'components/Settings/Registry/Stepper/CSVStepper.tsx',
   'components/Settings/Registry/Stepper/UrlStepper.tsx',
   'components/Settings/Registry/StyledTreeItem.tsx',
-  'components/SpacesSwitcher/DesignViewListItem.tsx',
-  'components/SpacesSwitcher/MainDesignsContent.tsx',
-  'components/SpacesSwitcher/MainViewsContent.tsx',
-  'components/SpacesSwitcher/MenuComponent.tsx',
-  'components/SpacesSwitcher/MobileViewSwitcher.tsx',
-  'components/SpacesSwitcher/MyDesignsContent.tsx',
-  'components/SpacesSwitcher/MyViewsContent.tsx',
-  'components/SpacesSwitcher/RecentContent.tsx',
-  'components/SpacesSwitcher/SharedContent.tsx',
-  'components/SpacesSwitcher/SpaceSwitcher.tsx',
-  'components/SpacesSwitcher/WorkspaceContent.tsx',
-  'components/SpacesSwitcher/WorkspaceModal.tsx',
-  'components/SpacesSwitcher/WorkspaceSwitcher.tsx',
-  'components/SpacesSwitcher/components.tsx',
+  'components/workspaces/SpacesSwitcher/DesignViewListItem.tsx',
+  'components/workspaces/SpacesSwitcher/MainDesignsContent.tsx',
+  'components/workspaces/SpacesSwitcher/MainViewsContent.tsx',
+  'components/workspaces/SpacesSwitcher/MenuComponent.tsx',
+  'components/workspaces/SpacesSwitcher/MobileViewSwitcher.tsx',
+  'components/workspaces/SpacesSwitcher/MyDesignsContent.tsx',
+  'components/workspaces/SpacesSwitcher/MyViewsContent.tsx',
+  'components/workspaces/SpacesSwitcher/RecentContent.tsx',
+  'components/workspaces/SpacesSwitcher/SharedContent.tsx',
+  'components/workspaces/SpacesSwitcher/SpaceSwitcher.tsx',
+  'components/workspaces/SpacesSwitcher/WorkspaceContent.tsx',
+  'components/workspaces/SpacesSwitcher/WorkspaceModal.tsx',
+  'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
+  'components/workspaces/SpacesSwitcher/components.tsx',
   'components/TroubleshootingComponent.tsx',
   'components/TypingFilter/index.tsx',
   'components/UserPreferences/index.tsx',
